+++ tests/integration.rs
@@ -29,10 +29,10 @@ use rand::{distributions::Standard, Rng};
 use safe_nd::{
     AData, ADataAddress, ADataAppendOperation, ADataEntry, ADataIndex, ADataOwner,
     ADataPermissions, ADataPubPermissionSet, ADataPubPermissions, ADataUnpubPermissionSet,
-    ADataUnpubPermissions, ADataUser, AppPermissions, AppendOnlyData, ClientFullId, Coins,
-    EntryError, Error as NdError, IData, IDataAddress, LoginPacket, MData, MDataAction,
-    MDataAddress, MDataEntries, MDataKind, MDataPermissionSet, MDataSeqEntryActions, MDataSeqValue,
-    MDataUnseqEntryActions, MDataValue, MDataValues, Message, MessageId, PubImmutableData,
+    ADataUnpubPermissions, ADataUser, AppPermissions, AppendOnlyData, ClientFullId, EntryError,
+    Error as NdError, IData, IDataAddress, LoginPacket, MData, MDataAction, MDataAddress,
+    MDataEntries, MDataKind, MDataPermissionSet, MDataSeqEntryActions, MDataSeqValue,
+    MDataUnseqEntryActions, MDataValue, MDataValues, Message, MessageId, Money, PubImmutableData,
     PubSeqAppendOnlyData, PubUnseqAppendOnlyData, PublicKey, Request, Response, Result as NdResult,
     SeqAppendOnly, SeqMutableData, Transaction, UnpubImmutableData, UnpubSeqAppendOnlyData,
     UnpubUnseqAppendOnlyData, UnseqAppendOnly, UnseqMutableData, XorName,
@@ -101,7 +101,7 @@ fn login_packets() {
     let login_packet_data = vec![0; 32];
     let login_packet_locator: XorName = env.rng().gen();
 
-    let balance = common::multiply_coins(COST_OF_PUT, 2);
+    let balance = common::multiply_money(COST_OF_PUT, 2);
     common::create_balance(&mut env, &mut client, None, balance);
 
     // Try to get a login packet that does not exist yet.
@@ -148,7 +148,7 @@ fn login_packets() {
         &mut env,
         &mut client,
         Request::GetBalance,
-        Coins::from_nano(1),
+        Money::from_nano(1),
     );
 
     // Getting login packet from non-owning client should fail.
@@ -212,7 +212,7 @@ fn create_login_packet_for_other() {
         &mut env,
         &mut established_client,
         Request::GetBalance,
-        Coins::from_nano(start_nano - nano_to_transfer),
+        Money::from_nano(start_nano - nano_to_transfer),
     );
     common::send_request_expect_ok(&mut env, &mut new_client, Request::GetBalance, COST_OF_PUT);
 
@@ -222,7 +222,7 @@ fn create_login_packet_for_other() {
         &mut established_client,
         Request::CreateLoginPacketFor {
             new_owner: *new_client.public_id().public_key(),
-            amount: Coins::from_nano(nano_to_transfer),
+            amount: Money::from_nano(nano_to_transfer),
             transaction_id: 2,
             new_login_packet: login_packet.clone(),
         },
@@ -234,7 +234,7 @@ fn create_login_packet_for_other() {
         &mut env,
         &mut established_client,
         Request::GetBalance,
-        Coins::from_nano(start_nano - nano_to_transfer),
+        Money::from_nano(start_nano - nano_to_transfer),
     );
 
     // Putting login packet to the same address with different balance should fail
@@ -259,7 +259,7 @@ fn create_login_packet_for_other() {
         &mut env,
         &mut established_client,
         Request::GetBalance,
-        Coins::from_nano(start_nano - 2 * nano_to_transfer),
+        Money::from_nano(start_nano - 2 * nano_to_transfer),
     );
 
     // Getting login packet from non-owning client should fail.
@@ -333,7 +333,7 @@ fn update_login_packet() {
 
 ////////////////////////////////////////////////////////////////////////////////
 //
-// Coins
+// Money
 //
 ////////////////////////////////////////////////////////////////////////////////
 
@@ -352,34 +352,34 @@ fn coin_operations() {
     );
 
     // Create A's balance
-    let amount_a = Coins::from_nano(10);
+    let amount_a = Money::from_nano(10);
     common::create_balance(&mut env, &mut client_a, None, amount_a);
     common::send_request_expect_ok(&mut env, &mut client_a, Request::GetBalance, amount_a);
 
-    let amount_b = Coins::from_nano(1);
+    let amount_b = Money::from_nano(1);
     common::create_balance(&mut env, &mut client_a, Some(&mut client_b), amount_b);
 
-    let amount_a = Coins::from_nano(8);
+    let amount_a = Money::from_nano(8);
     common::send_request_expect_ok(&mut env, &mut client_a, Request::GetBalance, amount_a);
     common::send_request_expect_ok(&mut env, &mut client_b, Request::GetBalance, amount_b);
 
-    // Transfer coins from A to B (first attempt with zero amount doesn't work)
-    let amount_zero = Coins::from_nano(0);
+    // Transfer Money from A to B (first attempt with zero amount doesn't work)
+    let amount_zero = Money::from_nano(0);
     let transaction_id = 2;
     common::send_request_expect_err(
         &mut env,
         &mut client_a,
-        Request::TransferCoins {
+        Request::TransferMoney {
             destination: *client_b.public_id().name(),
             amount: amount_zero,
             transaction_id,
         },
         NdError::InvalidOperation,
     );
-    common::transfer_coins(&mut env, &mut client_a, &mut client_b, 2, 3);
+    common::transfer_money(&mut env, &mut client_a, &mut client_b, 2, 3);
 
-    let amount_a = Coins::from_nano(6);
-    let amount_b = Coins::from_nano(3);
+    let amount_a = Money::from_nano(6);
+    let amount_b = Money::from_nano(3);
     common::send_request_expect_ok(&mut env, &mut client_a, Request::GetBalance, amount_a);
     common::send_request_expect_ok(&mut env, &mut client_b, Request::GetBalance, amount_b);
 }
@@ -394,19 +394,19 @@ fn create_balance_that_already_exists() {
     common::create_balance(&mut env, &mut client_a, None, 10);
     common::create_balance(&mut env, &mut client_a, Some(&mut client_b), 4);
 
-    let balance_a = Coins::from_nano(5);
-    let balance_b = Coins::from_nano(4);
+    let balance_a = Money::from_nano(5);
+    let balance_b = Money::from_nano(4);
 
     common::send_request_expect_ok(&mut env, &mut client_a, Request::GetBalance, balance_a);
     common::send_request_expect_ok(&mut env, &mut client_b, Request::GetBalance, balance_b);
 
     // Attempt to create the balance for B again. The request fails and A receives an error back.
     let transaction_id = 2;
-    let amount = Coins::from_nano(2);
+    let amount = Money::from_nano(2);
     common::send_request_expect_err(
         &mut env,
         &mut client_a,
-        Request::CreateBalance {
+        Request::CreateAccount {
             new_balance_owner: *client_b.public_id().public_key(),
             amount,
             transaction_id,
@@ -422,29 +422,29 @@ fn create_balance_that_already_exists() {
 
     // Attempt to create the balance for A again. This should however work for phase 1
     common::create_balance(&mut env, &mut client_a, None, 2);
-    let balance_a = Coins::from_nano(2);
+    let balance_a = Money::from_nano(2);
     common::send_request_expect_ok(&mut env, &mut client_a, Request::GetBalance, balance_a);
 }
 
 #[test]
-fn transfer_coins_to_balance_that_doesnt_exist() {
+fn transfer_money_to_balance_that_doesnt_exist() {
     let mut env = Environment::new();
 
     let mut client_a = env.new_connected_client();
     let client_b = env.new_connected_client();
 
-    let balance_a = Coins::from_nano(10);
+    let balance_a = Money::from_nano(10);
     common::create_balance(&mut env, &mut client_a, None, balance_a);
     common::send_request_expect_ok(&mut env, &mut client_a, Request::GetBalance, balance_a);
 
-    // Attempt transfer coins to B's balance which doesn't exist. The request fails and A receives
+    // Attempt transfer Money to B's balance which doesn't exist. The request fails and A receives
     // an error back.
     let transaction_id = 4;
-    let amount = Coins::from_nano(4);
+    let amount = Money::from_nano(4);
     common::send_request_expect_err(
         &mut env,
         &mut client_a,
-        Request::TransferCoins {
+        Request::TransferMoney {
             destination: *client_b.public_id().name(),
             amount,
             transaction_id,
@@ -468,7 +468,7 @@ fn coin_operations_by_app() {
     common::create_balance(&mut env, &mut client_a, None, 10);
 
     // Create an app with all permissions.
-    // The `perform_mutations` permission is required to send a `CreateBalance` request.
+    // The `perform_mutations` permission is required to send a `CreateAccount` request.
     let mut app = env.new_disconnected_app(client_a.public_id().clone());
     common::perform_mutation(
         &mut env,
@@ -477,7 +477,7 @@ fn coin_operations_by_app() {
             key: *app.public_id().public_key(),
             version: 1,
             permissions: AppPermissions {
-                transfer_coins: true,
+                transfer_money: true,
                 get_balance: true,
                 perform_mutations: true,
             },
@@ -490,29 +490,29 @@ fn coin_operations_by_app() {
         &mut env,
         &mut app,
         Request::GetBalance,
-        Coins::from_nano(10),
+        Money::from_nano(10),
     );
 
     // Create the destination client with balance.
     let mut client_b = env.new_connected_client();
     common::create_balance(&mut env, &mut client_b, None, 0);
 
-    // App transfers some coins.
+    // App transfers some Money.
     let transaction_id = 1;
-    common::transfer_coins(&mut env, &mut app, &mut client_b, 1, transaction_id);
+    common::transfer_money(&mut env, &mut app, &mut client_b, 1, transaction_id);
 
-    // Check the coins did actually transfer.
+    // Check the Money did actually transfer.
     common::send_request_expect_ok(
         &mut env,
         &mut client_a,
         Request::GetBalance,
-        Coins::from_nano(9),
+        Money::from_nano(9),
     );
     common::send_request_expect_ok(
         &mut env,
         &mut client_b,
         Request::GetBalance,
-        Coins::from_nano(1),
+        Money::from_nano(1),
     );
 }
 
@@ -522,10 +522,10 @@ fn coin_operations_by_app_with_insufficient_permissions() {
     let mut owner = env.new_connected_client();
 
     // Create initial balance.
-    let balance = Coins::from_nano(10);
+    let balance = Money::from_nano(10);
     common::create_balance(&mut env, &mut owner, None, balance);
 
-    // Create an app which does *not* have permission to transfer coins.
+    // Create an app which does *not* have permission to transfer Money.
     let mut app = env.new_disconnected_app(owner.public_id().clone());
     common::perform_mutation(
         &mut env,
@@ -535,7 +535,7 @@ fn coin_operations_by_app_with_insufficient_permissions() {
             version: 1,
             permissions: AppPermissions {
                 get_balance: false,
-                transfer_coins: false,
+                transfer_money: false,
                 perform_mutations: false,
             },
         },
@@ -550,15 +550,15 @@ fn coin_operations_by_app_with_insufficient_permissions() {
         NdError::AccessDenied,
     );
 
-    // The attempt to transfer some coins by the app fails.
+    // The attempt to transfer some Money by the app fails.
     let destination: XorName = env.rng().gen();
     let transaction_id = 1;
     common::send_request_expect_err(
         &mut env,
         &mut app,
-        Request::TransferCoins {
+        Request::TransferMoney {
             destination,
-            amount: Coins::from_nano(1),
+            amount: Money::from_nano(1),
             transaction_id,
         },
         NdError::AccessDenied,
@@ -731,7 +731,7 @@ fn put_append_only_data() {
         Request::PutAData(unpub_unseq_adata.clone()),
     );
 
-    let balance_a = Coins::from_nano(start_nano - 4);
+    let balance_a = Money::from_nano(start_nano - 4);
     common::send_request_expect_ok(&mut env, &mut client_a, Request::GetBalance, balance_a);
 
     // Get the data to verify
@@ -873,7 +873,7 @@ fn delete_append_only_data_that_doesnt_exist() {
         &mut env,
         &mut client,
         Request::GetBalance,
-        Coins::from_nano(start_nano),
+        Money::from_nano(start_nano),
     );
 }
 
@@ -1924,7 +1924,7 @@ fn put_immutable_data() {
         NdError::NoSuchBalance,
     );
 
-    // Create balances.  Client A starts with 2000 safecoins and spends 1000 to initialise
+    // Create balances.  Client A starts with 2000 safeMoney and spends 1000 to initialise
     // Client B's balance.
     let start_nano = 1_000_000_000_000;
     common::create_balance(&mut env, &mut client_a, None, start_nano * 2);
@@ -1938,8 +1938,8 @@ fn put_immutable_data() {
         NdError::InvalidOwners,
     );
 
-    let mut expected_a = Coins::from_nano(start_nano - 1);
-    let mut expected_b = Coins::from_nano(start_nano);
+    let mut expected_a = Money::from_nano(start_nano - 1);
+    let mut expected_b = Money::from_nano(start_nano);
     common::send_request_expect_ok(&mut env, &mut client_a, Request::GetBalance, expected_a);
 
     // Check they can both Put valid data.
@@ -2239,7 +2239,7 @@ fn auth_keys() {
 
     // Create an app with some permissions to mutate and view the balance.
     let permissions = AppPermissions {
-        transfer_coins: false,
+        transfer_money: false,
         perform_mutations: true,
         get_balance: true,
     };
@@ -2325,7 +2325,7 @@ fn auth_keys() {
 fn app_permissions() {
     let mut env = Environment::new();
     let mut owner = env.new_connected_client();
-    let balance = Coins::from_nano(1000);
+    let balance = Money::from_nano(1000);
     common::create_balance(&mut env, &mut owner, None, balance);
 
     // App 0 is authorized with permission to perform mutations.
@@ -2339,7 +2339,7 @@ fn app_permissions() {
             permissions: AppPermissions {
                 perform_mutations: true,
                 get_balance: false,
-                transfer_coins: false,
+                transfer_money: false,
             },
         },
     );
@@ -2354,7 +2354,7 @@ fn app_permissions() {
             key: *app_1.public_id().public_key(),
             version: 2,
             permissions: AppPermissions {
-                transfer_coins: false,
+                transfer_money: false,
                 get_balance: true,
                 perform_mutations: false,
             },
@@ -2365,7 +2365,7 @@ fn app_permissions() {
     // App 2 is not authorized.
     let mut app_2 = env.new_connected_app(owner.public_id().clone());
 
-    // App 3 is authorized with permission to transfer coins only.
+    // App 3 is authorized with permission to transfer Money only.
     let mut app_3 = env.new_disconnected_app(owner.public_id().clone());
     common::perform_mutation(
         &mut env,
@@ -2376,7 +2376,7 @@ fn app_permissions() {
             permissions: AppPermissions {
                 perform_mutations: false,
                 get_balance: false,
-                transfer_coins: true,
+                transfer_money: true,
             },
         },
     );
@@ -2448,7 +2448,7 @@ fn app_permissions() {
         NdError::AccessDenied,
     );
 
-    // Only the app with the transfer coins permission can perform mutable request.
+    // Only the app with the transfer Money permission can perform mutable request.
     for address in [pub_data_address, unpub_data_address].iter().cloned() {
         let append = ADataAppendOperation {
             address,
@@ -2479,17 +2479,17 @@ fn app_permissions() {
         );
     }
 
-    // A new client to credit coins to.
+    // A new client to credit Money to.
     let mut creditor = env.new_connected_client();
-    common::create_balance(&mut env, &mut creditor, None, Coins::from_nano(0));
+    common::create_balance(&mut env, &mut creditor, None, Money::from_nano(0));
 
-    // App 1 cannot transfer coins.
+    // App 1 cannot transfer Money.
     common::send_request_expect_err(
         &mut env,
         &mut app_1,
-        Request::TransferCoins {
+        Request::TransferMoney {
             destination: *creditor.public_id().name(),
-            amount: Coins::from_nano(50),
+            amount: Money::from_nano(50),
             transaction_id: 0,
         },
         NdError::AccessDenied,
@@ -2500,20 +2500,20 @@ fn app_permissions() {
         &mut env,
         &mut app_1,
         Request::GetBalance,
-        Response::GetBalance(Ok(Coins::from_nano(996))),
+        Response::GetBalance(Ok(Money::from_nano(996))),
     );
 
-    let amount = Coins::from_nano(900);
+    let amount = Money::from_nano(900);
     let expected = Response::Transaction(Ok(Transaction { id: 1, amount }));
     let name: XorName = env.rng().gen();
     let tag = 100;
     let data = SeqMutableData::new(name, tag, *owner.public_id().public_key());
 
-    // App 3 can transfer coins on behalf of the user
+    // App 3 can transfer Money on behalf of the user
     common::send_request_expect_ok(
         &mut env,
         &mut app_3,
-        Request::TransferCoins {
+        Request::TransferMoney {
             destination: *creditor.public_id().name(),
             amount,
             transaction_id: 1,
@@ -2669,7 +2669,7 @@ fn mutate_seq_mutable_data() {
     let mut env = Environment::new();
     let mut client = env.new_connected_client();
 
-    let balance = common::multiply_coins(COST_OF_PUT, 4);
+    let balance = common::multiply_money(COST_OF_PUT, 4);
     common::create_balance(&mut env, &mut client, None, balance);
 
     // Try to put sequenced Mutable Data.
@@ -2774,7 +2774,7 @@ fn mutate_unseq_mutable_data() {
     let mut env = Environment::new();
     let mut client = env.new_connected_client();
 
-    let balance = common::multiply_coins(COST_OF_PUT, 3);
+    let balance = common::multiply_money(COST_OF_PUT, 3);
     common::create_balance(&mut env, &mut client, None, balance);
 
     // Try to put unsequenced Mutable Data.
@@ -2866,8 +2866,8 @@ fn mutable_data_permissions() {
     let mut client_a = env.new_connected_client();
     let mut client_b = env.new_connected_client();
 
-    let balance_a = common::multiply_coins(COST_OF_PUT, 3);
-    let balance_b = common::multiply_coins(COST_OF_PUT, 3);
+    let balance_a = common::multiply_money(COST_OF_PUT, 3);
+    let balance_b = common::multiply_money(COST_OF_PUT, 3);
     common::create_balance(&mut env, &mut client_a, None, balance_a);
     common::create_balance(&mut env, &mut client_b, None, balance_b);
 
@@ -2948,7 +2948,7 @@ fn delete_mutable_data() {
     let mut client_a = env.new_connected_client();
     let mut client_b = env.new_connected_client();
 
-    let balance_a = common::multiply_coins(COST_OF_PUT, 3);
+    let balance_a = common::multiply_money(COST_OF_PUT, 3);
     common::create_balance(&mut env, &mut client_a, None, balance_a);
     common::create_balance(&mut env, &mut client_b, None, COST_OF_PUT);